# Conflicts:
#	onebase-module-bpm/onebase-module-bpm-build/src/main/java/com/cmsr/onebase/module/bpm/build/controller/engine/BpmEngineController.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowHisTaskDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowInstanceDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowNodeDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowSkipDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowTaskDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/FlowUserDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/dao/WarmDaoImpl.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowDefinition.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowHisTask.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowInstance.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowNode.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowSkip.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowTask.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/entity/FlowUser.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/repository/FlowInstanceRepository.java
#	onebase-module-bpm/onebase-module-bpm-engine-orm/src/main/java/com/cmsr/onebase/module/engine/orm/anyline/repository/FlowTaskRepository.java

+++ onebase-module-system/onebase-module-system-core/src/main/java/com/cmsr/onebase/module/system/service/dept/DeptServiceImpl.java
@@ -99,6 +99,11 @@ public class DeptServiceImpl implements DeptService {
         if (getChildDeptCount(id) > 0) {
             throw exception(DEPT_EXITS_CHILDREN);
         }
+        // 如果一个部门有用户，则不能删除
+        List<AdminUserDO> userListByDeptIds = adminUserService.getUserListByDeptIds(Collections.singleton(id));
+        if (CollectionUtils.isNotEmpty(userListByDeptIds)) {
+            throw exception(DEPT_DEL_FAILD_EXISTS_USERS);
+        }
         // 删除部门
         deptDataRepository.deleteById(id);
     }
@@ -173,6 +178,9 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public DeptDO getDept(Long id) {
+        if (id == null) {
+            return null;
+        }
         return deptDataRepository.findById(id);
     }
